04-Movie-Recommendation-System/requirements.txt
bleach==6.1.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dill==0.3.8
Flask==3.0.2
gunicorn==21.2.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
kaggle==1.6.6
MarkupSafe==2.1.5
numpy==1.26.4
packaging==23.2
pandas==2.2.0
python-dateutil==2.8.2
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
text-unidecode==1.3
threadpoolctl==3.3.0
tqdm==4.66.2
tzdata==2024.1
urllib3==2.2.1
webencodings==0.5.1
Werkzeug==3.0.1
